+++ simple_model/utils.py
@@ -1,3 +1,4 @@
+import inspect
 import re
 
 
@@ -21,3 +22,15 @@ def coerce_to_alpha(string: str) -> str:
 def snake_case(string: str) -> str:
     aux = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', string)
     return re.sub('([a-z0-9])([A-Z])', r'\1_\2', aux).lower()
+
+
+def is_not_special_object(obj):
+    return not any((
+        inspect.isclass(obj),
+        inspect.ismethod(obj),
+        inspect.isfunction(obj),
+        inspect.isgeneratorfunction(obj),
+        inspect.isgenerator(obj),
+        inspect.isroutine(obj),
+        isinstance(obj, property)
+    ))